+++ app/state.go
@@ -32,9 +32,12 @@ func (s State) Check() merkle.Tree {
 	return s.checkTx
 }
 
-// Commit stores the current Append() state as committed
-// starts new Append/Check state, and
-// returns the hash for the commit
+// Hash updates the tree
+func (s *State) Hash() []byte {
+	return s.deliverTx.Hash()
+}
+
+// Commit save persistent nodes to the database and re-copies the trees
 func (s *State) Commit() []byte {
 	var hash []byte
 	if s.persistent {