+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.2.3 (June 21, 2017)
+
+FEATURES:
+- [app] `Info()` now implements the ABCI handshake, allowing the app to recover from the latest state after crash
+- [app] `State.Hash()` returns the latest root of the DeliverTx tree
+- [iavl] `IAVLTree.BatchSet(key, value)` adds to the current batch in the NodeDB for atomic writes
+
+IMPROVEMENTS:
+- Better README
+- More comments in code 
+- Better testing from shell scripts
+
 ## 0.2.2 (June 5, 2017)
 
 BUG FIXES: